Refactor the class feature sheet
