+++ templates/item/parts/item-details.hbs
@@ -1,0 +1,5 @@
+<div class="tab details {{tab.cssClass}}" data-group="primary" data-tab="details">
+	{{#if featureTemplate}}
+		{{> (lookup this 'featureTemplate') }}
+	{{/if}}
+</div>
